Added crud operations to booking controller

+++ src/main/java/group4/backend/service/BookingService.java
@@ -1,7 +1,10 @@
 package group4.backend.service;
 
 import group4.backend.entities.Booking;
+import group4.backend.entities.Room;
+import group4.backend.entities.User;
 import group4.backend.repository.BookingRepository;
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +18,14 @@ public class BookingService {
     this.bookingRepository = bookingRepository;
   }
 
+  public Iterable<Booking> getAllBookings() {
+    return this.bookingRepository.findAll();
+  }
+
+  public Optional<Booking> getBooking(Integer id) {
+    return this.bookingRepository.findById(id);
+  }
+
   public boolean addBooking(Booking booking) {
     boolean added = false;
     if (canBeAdded(booking)) {
@@ -22,6 +33,10 @@ public class BookingService {
       added = true;
     }
     return added;
+  }
+
+  public Iterable<Booking> getAllBookingsByRoom(Room room) {
+    return this.bookingRepository.findByRoomProvider_Room(room);
   }
 
   private boolean canBeAdded(Booking booking) {
@@ -33,11 +48,22 @@ public class BookingService {
   }
 
   public boolean deleteBookingById(Integer id) {
-    boolean cannotBeFound = false;
-    this.bookingRepository.deleteById(id);
-    if (this.bookingRepository.existsById(id)) {
-      cannotBeFound = true;
+    boolean deleted = false;
+    Optional<Booking> bookingInDb = getBooking(id);
+    if (bookingInDb.isPresent()) {
+      this.bookingRepository.deleteById(id);
+      deleted = true;
     }
-    return cannotBeFound;
+    return deleted;
+  }
+
+  public boolean updateUser(Integer bookingId, Booking bookingToPut) {
+    boolean updated = false;
+    if (this.bookingRepository.findById(bookingId).isPresent()) {
+      this.bookingRepository.deleteById(bookingId);
+      this.bookingRepository.save(bookingToPut);
+      updated = true;
+    }
+    return updated;
   }
 }
